+++ CHANGELOG.md
@@ -13,6 +13,7 @@ https://github.com/gab1618/lapse/compare/latest..HEAD
 
 - List inline logs
 - See logs details
+- Lua script typings
 
 ## [0.3.0] - 2026-08-19